final-pjt-back/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
sqlparse==0.4.1
urllib3==1.26.7
